pyproject.toml
[tool.poetry]
name = "github-asana-bot"
version = "0.1.0"
description = ""
authors = ["karech"]

[tool.poetry.dependencies]
python = "^3.9"
github-webhooks-framework = "^0.1.7"
asana = "^0.10.3"
uvicorn = "^0.17.4"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
aioify = "^0.4.0"
sentry-sdk = "^1.5.5"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
isort = "^5.10.1"
black = "^22.1.0"
tox = "^3.24.5"
flake8 = "^4.0.1"
flake8-blind-except = "^0.2.0"
flake8-bugbear = "^22.1.11"
flake8-builtins = "^1.5.3"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.8.0"
flake8-debugger = "^4.0.0"
flake8-formatter-abspath = "^1.0.1"
flake8-isort = "^4.1.1"
flake8-logging-format = "^0.6.0"
flake8-mutable = "^1.2.0"
flake8-print = "^4.0.0"
flake8-quotes = "^3.3.1"
flake8-string-format = "^0.3.0"
flake8-broken-line = "^0.4.0"
flake8-eradicate = "^1.2.0"
flake8-multiline-containers = "^0.0.18"
flake8-use-fstring = "^1.3"
flake8-assertive = "^2.0.0"
flake8-tidy-imports = "^4.6.0"
ipython = "^8.0.1"

[tool.isort]
profile = "hug"
src_paths = ["bot"]
line_length = 120

[tool.black]
skip-string-normalization = true
line-length = 120
target-version = ["py39"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
